-Save team score and flag statuses when saving netdemo snapshots. -Revise and cleanup the netdemo control bindings.
SVN r2920 (0.6-dev)

+++ common/p_ctf.h
@@ -54,7 +54,7 @@ struct flagdata
 	AActor::AActorPtr actor;
 
 	// Integer representation of WHO has each flag (player id)
-	size_t flagger;
+	byte flagger;
 	int	pickup_time;
 
 	// Flag locations
